Check that a chromote connection is alive (#136)
* Check that a chromote connection is alive

When your computer goes to sleep, the websocket and all debugging sessions are closed. This PR ensures that Chromote and ChromoteSession report this accurately.

I distinguished between active and alive for Chromote but not ChromoteSession since the browser process continues to run in the background, even though we have lost our connection to it. There's no way to restore a closed debugging session. However, I'm not sure about the names; maybe I have alive and active around the wrong way?

Part of #94

* `devtools::document()` (GitHub Actions)

* Proof of concept for auto reconnect

* `devtools::document()` (GitHub Actions)

* Drop is_active_ private variable

* Better browser closing logic

* Track session/target separately

* Update news

* Polish docs

* Improve closing logic some more

* Use target_closed parameter

---------

Co-authored-by: hadley <hadley@users.noreply.github.com>