feat: Include path to chrome log file in failed to connect message
When we fail to launch chrome, we print out the contents of stderr in the error message. But there are often important details hidden in the stdout logs.

This PR adds the full path to the log file with the lines printed to stdout in the error message when chrome fails to launch.